Lesson 6 Исправление ошибок

+++ core-service/src/main/resources/db/migration/V1__init.sql
@@ -47,7 +47,7 @@ create table orders
 (
     id          bigserial primary key,
     username    varchar(255) not null,
-    total_price DOUBLE not null,
+    total_price decimal      not null,
     address     varchar(255),
     phone       varchar(255),
     created_at  timestamp default current_timestamp,
@@ -57,11 +57,11 @@ create table orders
 create table order_items
 (
     id                bigserial primary key,
-    product_id        bigint not null references products (id),
-    order_id          bigint not null references orders (id),
-    quantity          int    not null,
-    price_per_product int    not null,
-    price             int    not null,
+    product_id        bigint  not null references products (id),
+    order_id          bigint  not null references orders (id),
+    quantity          int     not null,
+    price_per_product decimal not null,
+    price             decimal not null,
     created_at        timestamp default current_timestamp,
     updated_at        timestamp default current_timestamp
 );
